Add correct reusable blocks support (Func or Eager)
A block that is used in Functional or Eager pipeline builder
twice will be replaced with `SharedProducer`, that will handle
`fit` method, and `SharedConsumer`, that will handle all
repeated uses.
Internally, `SharedProducer` passes the underlying block
to the `SharedConsumer`, making it possible to execute
pipelines with reusable blocks in concurrent environments.